scratch/requirements.txt

beautifulsoup4
pymongo
pyyaml
python-dateutil
